1) Updated OpenMP pragmas for CPU image processing code. 2) Fixed some more warnings.

+++ trunk/src/flitr/multi_cpuhistogram_consumer.cpp
@@ -39,7 +39,7 @@ void MultiCPUHistogramConsumerThread::run()
 
             if ((imageCount % imageStride)==0)
             {
-                int32_t imNum=0;
+                size_t imNum=0;
 
                 {
                 #pragma omp parallel for
